pyproject.toml
[tool.poetry]
name = "actions-example-python"
version = "0.1.0"
description = "An example of GH Actions with python"
authors = ["Brian Torres-Gil <brian@ixi.us>"]

[tool.poetry.dependencies]
python = "^3.6"
pan-os-python = "^1.0.1"
defusedxml = "^0.6.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
flake8 = "^3.8.3"
flake8-bugbear = "^20.1.4"
flake8-logging-format = "^0.6.0"
flake8-pep3101 = "^1.3.0"
flake8-builtins = "^1.5.3"
flake8-comprehensions = "^3.2.3"
flake8-string-format = "^0.3.0"
rope = "^0.17.0"
mypy = "^0.782"
bandit = "^1.6.2"
black = "^20.8b1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
